Revert "Generalize Respect.extend_dsl_with."
This reverts commit 751562006e6e83d56504871d6d0ce6a60a4f031f.

It is a good idea to generalize the effect of this method but
including the module in BaseDef is not a good idea. Some people
may want to extend the DSL by adding new context (inheriting from
BaseDef) not providing the basic commands and others.

A cleaner way would be to ask the users to include their module
in BasicCommands but it won't work due to the double inclusion
problem...

Thus, the best work-around is still to have this class enumeration.