Use my own #with_options instead of ActiveSupport one.
Option merger expects all methods send to it in the block to expect
an options hash as final argument. This is not the case here. #doc,
and FakeNameProxy#target does not. Also, it was not working as is
since we needed to wrap the receiver in FakeNameProxy before
to call 'with_options' on.